Player.java Changes: Created actionSelected method to return the result of the choice the user made in the runner class. Created isTurnOver method to return if a player is done with their turn. (This is used for the loop in the runner class)
MarioPartyGameRunner.java:
Changed the way the game essentially simulates. Before it was just a player rolls a die until they hit a star and, that's when they get to choose an answer. Now each turn is being simulated so the user can choose what they want to do in a turn. Their turn ends once a die is rolled.